+++ NEWS.md
@@ -19,6 +19,7 @@
 * Added `scale` and `df` arguments to `predict()` for `splm()` and `spautor()` models.
 * Add `dispersion` argument to `predict()` for `spglm()` and `spgautor()` models.
 * Enhanced numeric stability of deviance and pseudo R-squared for `spglm()` or `spgautor()` models when `family = "beta"`.
+* Added the `cov_type` argument to `covmatrix()` to return observed by observed, prediction by observed, observed by prediction, and prediction by prediction covariance matrices.
 
 
 # spmodel 0.7.0